refactor: clean up imports and formatting across multiple contracts
- Removed unused imports and organized import statements for clarity in various contracts, including BaseBridgeAdapter, LayerZeroAdapter, and StargateAdapter.
- Standardized formatting by adjusting whitespace and line breaks to enhance code readability.
- Updated the build-unit-test workflow by removing the lint step to streamline the CI process.
- Marked subproject commits for openzeppelin-contracts-upgradeable and pendle-core-v2-public as dirty to reflect local modifications.

+++ packages/chain-bridge/test/unit/CrossChainRegistryIntegration.t.sol
@@ -1,15 +1,16 @@
 // SPDX-License-Identifier: BUSL-1.1
 pragma solidity 0.8.28;
 
-import {Test} from "forge-std/Test.sol";
-import {CrossChainRegistry} from "../../src/contracts/CrossChainRegistry.sol";
-import {ProtocolAccessManager} from "@summerfi/access-contracts/contracts/ProtocolAccessManager.sol";
+import { CrossChainRegistry } from "../../src/contracts/CrossChainRegistry.sol";
+import { ProtocolAccessManager } from "@summerfi/access-contracts/contracts/ProtocolAccessManager.sol";
+import { Test } from "forge-std/Test.sol";
 
 /**
  * @title CrossChainRegistryIntegrationTest
  * @notice Integration test demonstrating the complete registry-based workflow
  */
 contract CrossChainRegistryIntegrationTest is Test {
+
     CrossChainRegistry public registry;
     ProtocolAccessManager public accessManager;
 
@@ -48,10 +49,7 @@ contract CrossChainRegistryIntegrationTest is Test {
 
         // Deploy registry
         vm.prank(governor);
-        registry = new CrossChainRegistry(
-            address(accessManager),
-            CURRENT_CHAIN_ID
-        );
+        registry = new CrossChainRegistry(address(accessManager), CURRENT_CHAIN_ID);
     }
 
     /*//////////////////////////////////////////////////////////////
@@ -65,46 +63,27 @@ contract CrossChainRegistryIntegrationTest is Test {
         // Register the relationship
         vm.expectEmit(true, true, true, true);
         emit CrossChainRelationshipRegistered(
-            arkAddress,
-            proxyAddress,
-            CURRENT_CHAIN_ID,
-            TARGET_CHAIN_ID,
-            ARK_FLEET_RELATIONSHIP
+            arkAddress, proxyAddress, CURRENT_CHAIN_ID, TARGET_CHAIN_ID, ARK_FLEET_RELATIONSHIP
         );
 
         vm.prank(governor);
         registry.registerCrossChainRelationship(
-            arkAddress,
-            proxyAddress,
-            CURRENT_CHAIN_ID,
-            TARGET_CHAIN_ID,
-            ARK_FLEET_RELATIONSHIP
+            arkAddress, proxyAddress, CURRENT_CHAIN_ID, TARGET_CHAIN_ID, ARK_FLEET_RELATIONSHIP
         );
 
         // Verify the relationship
-        (address targetContract, uint16 chainId) = registry.getTargetForSource(
-            arkAddress,
-            ARK_FLEET_RELATIONSHIP
-        );
+        (address targetContract, uint16 chainId) = registry.getTargetForSource(arkAddress, ARK_FLEET_RELATIONSHIP);
         assertEq(targetContract, proxyAddress);
         assertEq(chainId, TARGET_CHAIN_ID);
 
-        address sourceContract = registry.getSourceForTarget(
-            CURRENT_CHAIN_ID,
-            TARGET_CHAIN_ID,
-            proxyAddress,
-            ARK_FLEET_RELATIONSHIP
-        );
+        address sourceContract =
+            registry.getSourceForTarget(CURRENT_CHAIN_ID, TARGET_CHAIN_ID, proxyAddress, ARK_FLEET_RELATIONSHIP);
         assertEq(sourceContract, arkAddress);
 
         // Validate the relationship
         assertTrue(
             registry.isValidCrossChainPair(
-                arkAddress,
-                proxyAddress,
-                CURRENT_CHAIN_ID,
-                TARGET_CHAIN_ID,
-                ARK_FLEET_RELATIONSHIP
+                arkAddress, proxyAddress, CURRENT_CHAIN_ID, TARGET_CHAIN_ID, ARK_FLEET_RELATIONSHIP
             )
         );
     }
@@ -118,21 +97,13 @@ contract CrossChainRegistryIntegrationTest is Test {
 
         vm.prank(governor);
         registry.registerCrossChainRelationship(
-            arkAddress,
-            proxyAddress,
-            differentSourceChain,
-            differentTargetChain,
-            ARK_FLEET_RELATIONSHIP
+            arkAddress, proxyAddress, differentSourceChain, differentTargetChain, ARK_FLEET_RELATIONSHIP
         );
 
         // Verify with correct chain IDs
         assertTrue(
             registry.isValidCrossChainPair(
-                arkAddress,
-                proxyAddress,
-                differentSourceChain,
-                differentTargetChain,
-                ARK_FLEET_RELATIONSHIP
+                arkAddress, proxyAddress, differentSourceChain, differentTargetChain, ARK_FLEET_RELATIONSHIP
             )
         );
 
@@ -148,10 +119,7 @@ contract CrossChainRegistryIntegrationTest is Test {
         );
 
         // Verify target chain ID in relationship data
-        (address targetContract, uint16 chainId) = registry.getTargetForSource(
-            arkAddress,
-            ARK_FLEET_RELATIONSHIP
-        );
+        (address targetContract, uint16 chainId) = registry.getTargetForSource(arkAddress, ARK_FLEET_RELATIONSHIP);
         assertEq(targetContract, proxyAddress);
         assertEq(chainId, differentTargetChain);
     }
@@ -166,57 +134,23 @@ contract CrossChainRegistryIntegrationTest is Test {
 
         // Register ark with two different relationship types
         vm.prank(governor);
-        registry.registerCrossChainRelationship(
-            arkAddress,
-            proxy1,
-            CURRENT_CHAIN_ID,
-            TARGET_CHAIN_ID,
-            arkFleetType
-        );
+        registry.registerCrossChainRelationship(arkAddress, proxy1, CURRENT_CHAIN_ID, TARGET_CHAIN_ID, arkFleetType);
 
         vm.prank(governor);
-        registry.registerCrossChainRelationship(
-            arkAddress,
-            proxy2,
-            CURRENT_CHAIN_ID,
-            TARGET_CHAIN_ID,
-            bridgeType
-        );
+        registry.registerCrossChainRelationship(arkAddress, proxy2, CURRENT_CHAIN_ID, TARGET_CHAIN_ID, bridgeType);
 
         // Verify both relationships exist
-        (address target1, uint16 chain1) = registry.getTargetForSource(
-            arkAddress,
-            arkFleetType
-        );
+        (address target1, uint16 chain1) = registry.getTargetForSource(arkAddress, arkFleetType);
         assertEq(target1, proxy1);
         assertEq(chain1, TARGET_CHAIN_ID);
 
-        (address target2, uint16 chain2) = registry.getTargetForSource(
-            arkAddress,
-            bridgeType
-        );
+        (address target2, uint16 chain2) = registry.getTargetForSource(arkAddress, bridgeType);
         assertEq(target2, proxy2);
         assertEq(chain2, TARGET_CHAIN_ID);
 
         // Check both are valid
-        assertTrue(
-            registry.isValidCrossChainPair(
-                arkAddress,
-                proxy1,
-                CURRENT_CHAIN_ID,
-                TARGET_CHAIN_ID,
-                arkFleetType
-            )
-        );
-        assertTrue(
-            registry.isValidCrossChainPair(
-                arkAddress,
-                proxy2,
-                CURRENT_CHAIN_ID,
-                TARGET_CHAIN_ID,
-                bridgeType
-            )
-        );
+        assertTrue(registry.isValidCrossChainPair(arkAddress, proxy1, CURRENT_CHAIN_ID, TARGET_CHAIN_ID, arkFleetType));
+        assertTrue(registry.isValidCrossChainPair(arkAddress, proxy2, CURRENT_CHAIN_ID, TARGET_CHAIN_ID, bridgeType));
 
         // Check counts
         assertEq(registry.getRelationshipCount(arkFleetType), 1);
@@ -231,75 +165,36 @@ contract CrossChainRegistryIntegrationTest is Test {
 
         // Register two relationships
         vm.prank(governor);
-        registry.registerCrossChainRelationship(
-            ark1,
-            proxy1,
-            CURRENT_CHAIN_ID,
-            TARGET_CHAIN_ID,
-            ARK_FLEET_RELATIONSHIP
-        );
+        registry.registerCrossChainRelationship(ark1, proxy1, CURRENT_CHAIN_ID, TARGET_CHAIN_ID, ARK_FLEET_RELATIONSHIP);
 
         vm.prank(governor);
-        registry.registerCrossChainRelationship(
-            ark2,
-            proxy2,
-            CURRENT_CHAIN_ID,
-            TARGET_CHAIN_ID,
-            ARK_FLEET_RELATIONSHIP
-        );
+        registry.registerCrossChainRelationship(ark2, proxy2, CURRENT_CHAIN_ID, TARGET_CHAIN_ID, ARK_FLEET_RELATIONSHIP);
 
         assertEq(registry.getRelationshipCount(ARK_FLEET_RELATIONSHIP), 2);
 
         // Unregister first one
         vm.expectEmit(true, true, true, true);
-        emit CrossChainRelationshipUnregistered(
-            ark1,
-            proxy1,
-            CURRENT_CHAIN_ID,
-            TARGET_CHAIN_ID,
-            ARK_FLEET_RELATIONSHIP
-        );
+        emit CrossChainRelationshipUnregistered(ark1, proxy1, CURRENT_CHAIN_ID, TARGET_CHAIN_ID, ARK_FLEET_RELATIONSHIP);
 
         vm.prank(governor);
-        registry.unregisterCrossChainRelationship(
-            ark1,
-            ARK_FLEET_RELATIONSHIP,
-            TARGET_CHAIN_ID
-        );
+        registry.unregisterCrossChainRelationship(ark1, ARK_FLEET_RELATIONSHIP, TARGET_CHAIN_ID);
 
         // Verify first is gone, second remains
         assertEq(registry.getRelationshipCount(ARK_FLEET_RELATIONSHIP), 1);
-        assertFalse(
-            registry.isSourceContractRegistered(ark1, ARK_FLEET_RELATIONSHIP)
-        );
-        assertTrue(
-            registry.isSourceContractRegistered(ark2, ARK_FLEET_RELATIONSHIP)
-        );
+        assertFalse(registry.isSourceContractRegistered(ark1, ARK_FLEET_RELATIONSHIP));
+        assertTrue(registry.isSourceContractRegistered(ark2, ARK_FLEET_RELATIONSHIP));
 
         // Verify second relationship still works
-        (address targetContract, uint16 chainId) = registry.getTargetForSource(
-            ark2,
-            ARK_FLEET_RELATIONSHIP
-        );
+        (address targetContract, uint16 chainId) = registry.getTargetForSource(ark2, ARK_FLEET_RELATIONSHIP);
         assertEq(targetContract, proxy2);
         assertEq(chainId, TARGET_CHAIN_ID);
 
-        address sourceContract = registry.getSourceForTarget(
-            CURRENT_CHAIN_ID,
-            TARGET_CHAIN_ID,
-            proxy2,
-            ARK_FLEET_RELATIONSHIP
-        );
+        address sourceContract =
+            registry.getSourceForTarget(CURRENT_CHAIN_ID, TARGET_CHAIN_ID, proxy2, ARK_FLEET_RELATIONSHIP);
         assertEq(sourceContract, ark2);
 
         assertTrue(
-            registry.isValidCrossChainPair(
-                ark2,
-                proxy2,
-                CURRENT_CHAIN_ID,
-                TARGET_CHAIN_ID,
-                ARK_FLEET_RELATIONSHIP
-            )
+            registry.isValidCrossChainPair(ark2, proxy2, CURRENT_CHAIN_ID, TARGET_CHAIN_ID, ARK_FLEET_RELATIONSHIP)
         );
     }
 
@@ -312,54 +207,22 @@ contract CrossChainRegistryIntegrationTest is Test {
 
         // Register same source-target pair with different types
         vm.prank(governor);
-        registry.registerCrossChainRelationship(
-            arkAddress,
-            proxyAddress,
-            CURRENT_CHAIN_ID,
-            TARGET_CHAIN_ID,
-            type1
-        );
+        registry.registerCrossChainRelationship(arkAddress, proxyAddress, CURRENT_CHAIN_ID, TARGET_CHAIN_ID, type1);
 
         vm.prank(governor);
-        registry.registerCrossChainRelationship(
-            arkAddress,
-            proxyAddress,
-            CURRENT_CHAIN_ID,
-            TARGET_CHAIN_ID,
-            type2
-        );
+        registry.registerCrossChainRelationship(arkAddress, proxyAddress, CURRENT_CHAIN_ID, TARGET_CHAIN_ID, type2);
 
         // Both should exist independently
         assertTrue(registry.isSourceContractRegistered(arkAddress, type1));
         assertTrue(registry.isSourceContractRegistered(arkAddress, type2));
 
         // Both should remain active since status update functionality was removed
-        assertTrue(
-            registry.isValidCrossChainPair(
-                arkAddress,
-                proxyAddress,
-                CURRENT_CHAIN_ID,
-                TARGET_CHAIN_ID,
-                type1
-            )
-        );
-        assertTrue(
-            registry.isValidCrossChainPair(
-                arkAddress,
-                proxyAddress,
-                CURRENT_CHAIN_ID,
-                TARGET_CHAIN_ID,
-                type2
-            )
-        );
+        assertTrue(registry.isValidCrossChainPair(arkAddress, proxyAddress, CURRENT_CHAIN_ID, TARGET_CHAIN_ID, type1));
+        assertTrue(registry.isValidCrossChainPair(arkAddress, proxyAddress, CURRENT_CHAIN_ID, TARGET_CHAIN_ID, type2));
 
         // Unregister one type
         vm.prank(governor);
-        registry.unregisterCrossChainRelationship(
-            arkAddress,
-            type1,
-            TARGET_CHAIN_ID
-        );
+        registry.unregisterCrossChainRelationship(arkAddress, type1, TARGET_CHAIN_ID);
 
         // Should only affect the specific type
         assertFalse(registry.isSourceContractRegistered(arkAddress, type1));
@@ -371,8 +234,7 @@ contract CrossChainRegistryIntegrationTest is Test {
 
     function test_supportedRelationshipTypes() public {
         // Initially should have ARK_FLEET type pre-registered in constructor
-        bytes32[] memory supportedTypes = registry
-            .getSupportedRelationshipTypes();
+        bytes32[] memory supportedTypes = registry.getSupportedRelationshipTypes();
         assertEq(supportedTypes.length, 2);
         assertEq(supportedTypes[0], keccak256("ADAPTER_PEER"));
         assertEq(supportedTypes[1], keccak256("ARK_FLEET"));
@@ -383,13 +245,7 @@ contract CrossChainRegistryIntegrationTest is Test {
         address proxyAddress = makeAddr("proxy");
 
         vm.prank(governor);
-        registry.registerCrossChainRelationship(
-            arkAddress,
-            proxyAddress,
-            CURRENT_CHAIN_ID,
-            TARGET_CHAIN_ID,
-            newType
-        );
+        registry.registerCrossChainRelationship(arkAddress, proxyAddress, CURRENT_CHAIN_ID, TARGET_CHAIN_ID, newType);
 
         // Should now have 2 supported types
         supportedTypes = registry.getSupportedRelationshipTypes();
@@ -398,11 +254,12 @@ contract CrossChainRegistryIntegrationTest is Test {
         // Check both types are present (order may vary)
         bool hasArkFleet = false;
         bool hasNewType = false;
-        for (uint i = 0; i < supportedTypes.length; i++) {
+        for (uint256 i = 0; i < supportedTypes.length; i++) {
             if (supportedTypes[i] == keccak256("ARK_FLEET")) hasArkFleet = true;
             if (supportedTypes[i] == newType) hasNewType = true;
         }
         assertTrue(hasArkFleet);
         assertTrue(hasNewType);
     }
+
 }
